% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discogger-package.R
\docType{package}
\name{discogger-package}
\alias{discogger}
\alias{discogger-package}
\title{discogger: Interface to the 'Discogs' API}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Provides tools to interact with the 'Discogs' API <https://www.discogs.com/developers/#>.
}
\author{
\strong{Maintainer}: Ewen Henderson \email{ewenhenderson@gmail.com}

}
\keyword{internal}
