% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{discogs_label_releases}
\alias{discogs_label_releases}
\title{Get metadata for a Discogs Label's Releases}
\usage{
discogs_label_releases(label_id, access_token = discogs_api_token())
}
\arguments{
\item{label_id}{The ID of the Label.}

\item{access_token}{Discogs personal access token, defaults to \code{discogs_api_token}.}
}
\value{
a \code{discogs_label_releases} object that contains the extracted content from the request,
the original JSON response object and the request path.
}
\description{
Return tidy metadata for a Label's (a label, company, recording studio, location,
or other entity involved with Artists and Releases) Releases listed on Discogs.
}
\examples{
\dontrun{
discogs_label_releases(label_id = 314)
}
}
