% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplace.R
\name{discogs_mkt_price_suggest}
\alias{discogs_mkt_price_suggest}
\title{Get price suggestions for a Discogs release}
\usage{
discogs_mkt_price_suggest(release_id, access_token = discogs_api_token())
}
\arguments{
\item{release_id}{The ID of the Release.}
}
\value{
a \code{discogs_database} object that contains the extracted content from the request,
the original JSON response object and the request path.
}
\description{
Return price suggestions for a release (a particular physical or digital object
released by one or more artists) listed on Discogs.
}
\examples{
\dontrun{
discogs_mkt_price_suggest(release_id = 240007)
}
}
