% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limiting.R
\name{discogs_rate_limit}
\alias{discogs_rate_limit}
\title{Check Discogs API user rate limits}
\usage{
discogs_rate_limit(access_token = discogs_api_token())
}
\arguments{
\item{access_token}{Discogs personal access token, defaults to \code{discogs_api_token}.}
}
\value{
None (invisible NULL)
}
\description{
This function prints a users Discogs API rate limit information. All calls to the API
deduct one unit from rate limits. Discogs rate limiting tracks your requests using a moving
average over a 60 second window. If no requests are made in 60 seconds, your window will reset.
For more information, check the \href{https://www.discogs.com/developers/#page:home,header:home-rate-limiting}{rate limiting API documentation}
}
\examples{
\dontrun{
discogs_rate_limit()
}
}
