% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{discogs_release}
\alias{discogs_release}
\title{Get metadata for a Discogs Release}
\usage{
discogs_release(release_id, mkt_currency = c("GBP", "USD", "EUR", "CAD",
  "AUD", "JPY", "CHF", "MXN", "BRL", "NZD", "SEK", "ZAR"),
  access_token = discogs_api_token())
}
\arguments{
\item{release_id}{The ID of the Release.}

\item{mkt_currency}{Currency for marketplace data. Defaults to GBP (must be one of:
"GBP", "USD", "EUR", "CAD", "AUD", "JPY", "CHF", "MXN", "BRL", "NZD", "SEK", "ZAR").}

\item{access_token}{Discogs personal access token, defaults to \code{discogs_api_token}.}
}
\value{
a \code{discogs_database} object that contains the extracted content from the request,
the original JSON response object and the request path.
}
\description{
Return metadata for a Release (a particular physical or digital object
released by one or more Artists) listed on Discogs.
}
\examples{
\dontrun{
discogs_release(release_id = 240007)
}
}
