% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{discogs_release_master}
\alias{discogs_release_master}
\title{Get metadata for a Discogs master release}
\usage{
discogs_release_master(master_id, access_token = discogs_api_token())
}
\arguments{
\item{master_id}{The Master ID of a release}

\item{access_token}{Discogs personal access token, defaults to \code{discogs_api_token}.}
}
\value{
a \code{discogs_database} object that contains the extracted content from the request,
the original JSON response object and the request path.
}
\description{
Return metadata for a master release (a set of similar Releases) listed
on Discogs.
}
\examples{
\dontrun{
discogs_release_master(master_id = 684265)
}
}
