% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{discogs_search}
\alias{discogs_search}
\title{Search Discogs database}
\usage{
discogs_search(params, n_results = NULL,
  access_token = discogs_api_token())
}
\arguments{
\item{params}{named list of parameters
(see \href{https://www.discogs.com/developers/#page:database,header:database-search}{SEARCH API docs}
for available parameters)}

\item{n_results}{(optional) set limit on no. of results (numeric; NULL by default)}

\item{access_token}{Discogs personal access token, defaults to \code{discogs_api_token}.}
}
\description{
Issue a search query to the Discogs database using any number of
(optional) parameters.
}
\examples{
\dontrun{
discogs_search(params = list(release_title = "Purple Rain",
artist = "Prince"), n_results = 10)
}

}
