% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{discogs_token}
\alias{discogs_token}
\title{Get or set DISCOGS_API_TOKEN value}
\usage{
discogs_token(force = FALSE)
}
\arguments{
\item{force}{force setting a new Discogs API token for the current environment?}
}
\value{
atomic character vector containing the Discogs API token
}
\description{
The API wrapper functions in this package all rely on a Discogs access token
residing in the environment variable \code{DISCOGS_API_TOKEN}. The
easiest way to accomplish this is to set it in the `\code{.Renviron}` file in your
home directory.
}
