% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{discogs_user_collection}
\alias{discogs_user_collection}
\title{Get metadata for a Discogs User Collection}
\usage{
discogs_user_collection(user_name, folder_id = 0,
  access_token = discogs_api_token())
}
\arguments{
\item{user_name}{The username of the Collection you are trying to request.}

\item{folder_id}{The ID of the Collection folder (default value is 0,
the “All” folder).}

\item{access_token}{Discogs personal access token, defaults to \code{discogs_api_token}.}
}
\value{
a \code{discogs_collection} object that contains the extracted content from the request,
the original JSON response object and the request path.
}
\description{
Return metadata for releases in a Discogs User Collection.
}
\examples{
\dontrun{
discogs_user_collection(user_name = "rodneyfool")
}
}
