% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist.R
\name{get_discogs_artist_releases}
\alias{get_discogs_artist_releases}
\title{Get metadata for a Discogs Artist's Releases}
\usage{
get_discogs_artist_releases(artist_id, access_token = discogs_api_token())
}
\arguments{
\item{artist_id}{The ID of the Artist.}

\item{access_token}{Discogs personal access token, defaults to \code{discogs_api_token}.}
}
\value{
a tibble
}
\description{
Return tidy metadata for an Artist's (a person who contributed
to a Release, in some capacity) Releases listed on Discogs.
}
