% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{get_discogs_user_collection}
\alias{get_discogs_user_collection}
\title{Get metadata for a user's Discogs Collection}
\usage{
get_discogs_user_collection(user_name, folder_id = 0,
  access_token = discogs_api_token())
}
\arguments{
\item{user_name}{The username of the Collection you are trying to request.}

\item{folder_id}{The ID of the Collection folder (default value is 0,
the “All” folder).}

\item{access_token}{Discogs personal access token, defaults to \code{discogs_api_token}.}
}
\value{
a tibble
}
\description{
Return tidy metadata for releases in a user's Discogs collection.
}
