% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.discogs_collection}
\alias{print.discogs_collection}
\title{Slightly more human-readable output for discogs_collection objects}
\usage{
\method{print}{discogs_collection}(x, ...)
}
\arguments{
\item{x}{a discogs_collection object}

\item{...}{ignored}
}
\description{
Slightly more human-readable output for discogs_collection objects
}
