% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.discogs_database}
\alias{print.discogs_database}
\title{Slightly more human-readable output for discogs_database objects}
\usage{
\method{print}{discogs_database}(x, ...)
}
\arguments{
\item{x}{a discogs_database object}

\item{...}{ignored}
}
\description{
Slightly more human-readable output for discogs_database objects
}
