% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{get_discogs_release}
\alias{get_discogs_release}
\title{Get metadata for a Discogs release}
\usage{
get_discogs_release(release_id, access_token = discogs_api_token())
}
\arguments{
\item{release_id}{The ID of the release.}

\item{access_token}{Discogs personal access token, defaults to \code{discogs_api_token}.}
}
\value{
a tibble
}
\description{
Return tidy metadata for a release (a particular physical or digital object
released by one or more Artists) listed on Discogs.
}
